Calculate Means

calc.means calculates various types of central-tendency measures for the vector passed to argument a

calc.means(
  data,
  mean.type = "am",
  post.op = eval,
  as.zscore = FALSE,
  use.population = FALSE,
  ...
)

Arguments

data

(numeric): A numeric vector

mean.type

(string): A character vector of the types of mean value operations to execute (see "Details")

post.op

(call | eval): A function that will process the output before returning

as.zscore

(logical | FALSE) Should the output be transformed to Z-scores?

use.population

(logical | FALSE) Should the population standard deviation be used (ignored when as.zscore==FALSE)?

...

Additional arguments passed to other functions

Value

Arguments data, mean.type, and post.op determine the return type.

Details

If mean.type contains multiple values, a list object is returned containing one member for each output; otherwise, a single value is returned with a class dependent on the value of mean.type. Valid mean.type values are as follows:

"am" (Arithmetic-Mean https://en.wikipedia.org/wiki/Arithmetic_mean): Return the sum of vector values divided by the magnitude of the vector

"zm" (Zero-Mean): Return the vector with the mean of the vector subtracted from all values

"gm" (Geometric Mean https://en.wikipedia.org/wiki/Geometric_mean): Return the n-th root of the product of an n-element input vector. Negative input values are allowed, resulting in a complex result

"hm" (Harmonic Mean https://en.wikipedia.org/wiki/Harmonic_mean): Return the reciprocal of the arithmetic mean of the reciprocal of the elements of the input vector

"rms"(Root Mean Squared https://en.wikipedia.org/wiki/Root_mean_square): Return the square-root of the arithmetic mean of the squares of the elements of the input vector

See also