Execute a Function as Recursively

call.recursion Executes a recursive calls to fun based on the output of test up to max.iter times

Usage

call.recursion(
  x,
  fun,
  test,
  ...,
  nxt = I,
  max.iter = 1,
  cur.iter = 0,
  simplify = TRUE
)

Arguments

x

The input object

fun

A function that operates on x and produces output

test

A single-argument function returning a single Boolean: FALSE stops iteration

...

Additional arguments passed to fun

nxt

A function that operates on the current output of fun(x) to send to the next iterative call to fun()

max.iter

(integer) The maximum number of iterations

cur.iter

(integer) The current iteration index

simplify

(logical) Should only the last value be returned (TRUE) or intermediate values as well (FALSE)?

See also