Import keyring Entries

keyring_import registers exported keyrings (see keyring_export)

Usage

keyring_import(
  kr_data,
  kr_name = names(kr_data),
  kr_pass = NULL,
  dry.run = FALSE
)

Arguments

kr_data

The named list of exported keyring data: each name should be the name of an existing keyring.

kr_name

(string) The names of the target keyring(s) to populate: defaults to the names of kr_data

kr_pass

(string) The password to unlock existing keyrings:

A named list indicating the per-keyring mapping

A single password (used for all supplied keyrings)

An unnamed list the same length as kr_name

dry.run

(logical|FALSE) When TRUE, the expression that would be evaluated is returned (passwords are redacted)

Value

A list of logical vectors indicating success or failure of import for each keyring entry

See also