Store an encrypted keyring key

The kr_key class

Slots

service

See key_get

username

See key_get

keyring

See key_get

key,password

An anonymous function to retrieve and return the value of the current keyring key.

update

An anonymous function to set a new value for the keyring key referenced by the object.

See also