Create an Entry into a Development Log

log_note serves as a wrapper for cat and provides an easy way to write to a plain-text file Even though the original use-case is the real-time capturing of developer notes, it can be used for incremental writing to a file when needed.

Usage

log_note(..., file = "", show = FALSE, append = TRUE)

Arguments

...

(dots_list): The string(s) of text to write to the file. Multiple values are collapsed into a single string separated by a newline. Expressions are allowed and will have the captured output included.

file

(string|"") The name of the file to (over)write. If no file is given, the output is sent to the console.

show

(logical | FALSE) When TRUE, print the contents of file (requires library readtext).

append

(logical | TRUE) When TRUE new content is appended.

Value

See parameter show