Calculate percentages from odds ratios

odds2probs converts odds ratios to percentages.

Usage

odds2probs(..., log = FALSE)

Arguments

...

(dots_list): Numeric vectors or "odds" strings (e.g., "x:y"). For vectors or string representations, the number of operands can be of any length but must represent positive real values.

log

(logical) Are the inputs log-odds? If TRUE, arguments of length-1 are processed as-is; otherwise, arguments are natural-logged.

Value

A numeric list length as the input(s) representing the percentage of the odds ratio(s): values that cannot be converted will return NULL.

Note

Length-1 arguments are assumed to be raw odds (or log-odds)

See also

Examples

odds2probs(c(3, 4), c(3L, 4L), "3:4", c("3", "4"), c(4,5,6), "6:3:1", c(3,-1), "1:1", 1, 2/5, "2:3", 2)