Radix Conversion

radix converts the representation of the input(s) using a radix basis. This is not to be confused with a conversion between scales.

Usage

radix(x, ...)

Arguments

x

(numeric[],symbol[]) One or more whole numbers: coercion for non-numeric and fractional input will result in loss of precision. For hexadecimal values, be sure to use the "0x" prefix.

...

(string,symbol,integer) The target radix to which the values in x will be converted.

Length: If two values are provided, the output is the conversion from the first radix to the second

String/symbol: If given as a strings or symbols, the following are supported (partial matching supported):

binary

octal

decimal

duodecimal

hexadecimal

vigesimal

sexagesimal

Value

The decimal representation of the input(s) using the supplied radix.

See also

Examples

# As binary
radix(12, bin)
radix(c(12, 20), bin)

# To hexadecimal
radix(c(12, 20), hex)

# As hex to decimal
radix(c(18, 32), hex, dec)

# Hex to binary vs. as hex to binary (equivalent outcomes)
radix(c(18, 32), hex, bin)
radix(!!radix(c(12, 20), hex), hex, bin)

# Arbitrary radix
radix(c(10, 26), 3)

# Arbitrary radix to hex
radix(c(10, 26), 3, h)

# Mixed input types to binary
radix(c(100, "0110"), b)
radix(c(100, "1001", 0x1D), b)