Ratio Calculator

ratio calculates one of the following ratio types:

"of.sum" (relative to the sum of i)

"of.max" (relative to maximum value)

"cumulative" (cumulative total vs. total)

.
Using the related operator %ratio% assumes simple division by the total of i.

Usage

ratio(
  i,
  type = "of.sum",
  as_density = FALSE,
  sort.type = c("none", "value", "label", "num_label"),
  ...
)

Arguments

i

(vector) numeric vector

type

(string[]) One or more types of ratio methods to use (see Details): a vector of supported values is supported

as_density

(logical) TRUE returns x * p(1 - p), where p is a vector of cumulative proportions of x

sort.type

(string,function) One of the following:

"none" to return along the original order (default)

"value" to sort by value

"label" to sort by name: all elements must be named

"num_label" to sort by numeral names: all elements must be named

...

dots_list: provided for backwards-compatibility as the function signature changes. Sorting occurs before the input is further processed.

Value

A numeric vector

Note

Because this function produces values on a ratio scale, all values are internally shifted such that all values are >= 0.

See also

Examples

 (x <- sample(-100:100, 30))

 (d <- cbind(
   x
   , of.sum = ratio(x, of.sum)
   , of.sum.dens = ratio(x, of.sum, as = TRUE)
   , of.max = ratio(x, of.max)
   , of.max.dens = ratio(x, of.max, as = TRUE)
   , cumulative = ratio(x, cumulative)
   , cumulative.dens = ratio(x, cumulative, as = TRUE)
   ))

 plot(of.sum ~ x, data = d, main = "of.sum", col = "red")
 plot(of.max ~ x, data = d, main = "of.max", col = "green")
 plot(cumulative ~ x, data = d, main = "cumulative", col = "blue")
 plot(of.sum.dens ~ x, data = d, main = "of.sum w/ density", col = "red")
 points(of.max.dens ~ x, data = d, main = "of.max w/ density", col = "green")
 points(cumulative.dens ~ x, data = d, main = "cumulative w/ density", col = "blue")

 # sort.type:
 (x2 <- sample(100, 5, TRUE) %>% rlang::set_names(. + 50))

 list(
   `orig := table(...)` = x2
   , `ratio(<args>, sort.type = none)` = ratio(x2, sort.type = "none")
   , `ratio(<args>, sort.type = value)` = ratio(x2, sort.type = "value")
   , `ratio(<args>, sort.type = label)` = ratio(x2, sort.type = "label")
   , `ratio(<args>, sort.type = num_label)` = ratio(x2, sort.type = "num_label")
   )