"Between" Operator

test_between and related operator %><% chunks out each of the inputs using row-wise 'apply' and returns a Boolean result matrix. When arguments a and z are vectors or tensors, "between" is interpreted as the existence of the intersection of the domains spanned by the vectors/tensors. When temporal is TRUE, an ordinal response is returned depending on the value of as.text

Usage

test_between(a, z, temporal = FALSE, as.text = FALSE)

Arguments

a

Preferably a data table or data frame, but any object that can be coerced into such

z

Preferably a data table or data frame, but any object that can be coerced into such

temporal

(logical) Should relative temporal output (e.g., before, during, after) be returned as integers -1 0 1 rather than logical values?

as.text

(logical) Should relative temporal output (e.g., before, during, after) be returned as words rather than logical values?

Value

An Boolean object comprised of the results of testing for "between-ness"

Details

as.text FALSE: the return matrix is comprised of "-1, 0, 1" denoting "before" "during" "after"

as.text TRUE: the return matrix is comprised of "before" "during" "after"

See also