Vectorized Logical Tests

vlike is a vectorized version of like allowing a pattern vector to be supplied

Usage

vlogical(
  vector,
  vpattern,
  test = stringi::stri_detect_regex,
  simplify_with = NULL,
  ...
)

Arguments

vector

A vector or dimensional object to be processed (e.g., matrix, data.frame, etc.)

vpattern

A named vector of patterns to be matched

test

(function) The function to use for logical testing: the function should be appropriate for the values of vpattern accepting vector as the first argument, the pattern as the second, and ... as the third (even if not used).

simplify_with

(function) A function to column-wise simplify the logical matrix when not empty

...

Additional arguments to be sent to the function held by argument test

Value

A logical matrix, with rows of the same length as vector and columns the length of vpattern when simplify_with is not provided. Otherwise, the result of applying simplify_with on each column.

See also

Examples

  # Test 1:
vlogical(
  vector = letters
  , vpattern = c(sample(LETTERS, 5), sample(letters, 5))
  , simplify_with = mean
  , ignore.case = TRUE
  )

  # Test 2:
  vl_data <- sapply(1:100000, function(i){
        c(a = sample(LETTERS, 1), b = sample(letters, 1))
      }) |>
    data.table::as.data.table() |>
  t();

head(vl_data, 10);
dim(vl_data);

vlogical(
  vector = vl_data
  , vpattern = c(sample(LETTERS, 5), sample(letters, 5)) |> rlang::set_names()
  , test = \(vector, pattern, ...){ paste(unique(pattern), collapse = "|") |> grepl(vector)}
  , simplify_with = NULL
  , ignore.case = TRUE
  )