Execute a Function as Recursively

call.recursion Executes a recursive calls to fun based on the output of test up to max.iter times

call.recursion(x, fun, test, nxt, max.iter = 1, cur.iter = 0, simplify = TRUE)

Arguments

x

The input object

fun

A function that operates on x and produces output

test

A single-argument function returning a single Boolean: FALSE stops iteration

nxt

A function that operates on the current output of fun(x) to send to the next iterative call to fun()

max.iter

(integer) The maximum number of iterations

cur.iter

(integer) The current iteration index

simplify

(logical) Should only the last value be returned (TRUE) or intermediate values as well (FALSE)?

Examples

book.of.utilities::call.recursion(
x = sample(1000, size = 100)
, fun = \(x){ abs(x - mean(x, na.rm = TRUE)) }
, test = \(x){
    if (length(unique(x)) == 1 || is.na(sd(x))){
      FALSE
    } else {
      i <- abs(x - mean(x, na.rm = TRUE))/sd(x);
      mean(i > 3) <= 0.7
    }
  }
, nxt = \(x){ sample(x, size = length(x)- 1, prob = runif(n = length(x), min = 0.1, max = 0.9)) }
, max.iter = 1000
, simplify = !FALSE
)
#> [[1]]
#> [1] 1.198539e-09 1.198539e-09
#>