Correlated IF-THEN-ELSE Comparison

For each element E in cond, when E is TRUE, the corresponding index of result$true is returned: the same for cond == FALSE. The important thing to remember is that the function returns a value for TRUE and FALSE per row.
For result, the easiest way to set the available choices is to use %tf%; otherwise, a environment(-like) object with members named true and false.

cond %??% result

Arguments

cond

(logical) A rank-n object for which each row contains TRUE or FALSE

result

(vector) A rank-n object for which each row contains resultant values for each element of cond[m, n] containing values for TRUE and FALSE. The result can be of a different structure than the input as is the case when each result is of heterogeneous lengths.

Value

A list of values, one element per row of input.

See also