Ranking Algorithm

ranking.algorithm ranks values across a vector of values

ranking.algorithm(
  scores,
  rank.size,
  poss.scores,
  test = `>=`,
  rescale = 1,
  map.only = FALSE,
  show.all = FALSE,
  use.futures = FALSE,
  dflt.fn = min,
  map.fn = (purrr::map)
)

Arguments

scores

(vector) The vector of numeric scores to rank

rank.size

(integer) The number of ranks to allow

poss.scores

(vector) The distinct values demarcating each rank threshold

test

(object) A function or operator that takes in as input the scores, performs a conditional test, and returns a logical vector of equal length

rescale

(numeric) A multiplier that scales the scores vector up or down

map.only

(logical|FALSE) When TRUE, a data.table object of the rank and rank thresholds is returned

show.all

(logical|FALSE) When TRUE, a data.table object of the scores, ranks thresholds, and ranks are returned

use.futures

(logical|FALSE) When TRUE, the operation is parallelized using futures

dflt.fn

(function) The default function to use when replacing NA values in ranks

map.fn

(function) The function to iterate the scores when computing the output

See also