Vectorized Logical Tests

vlike is a vectorized version of like allowing a pattern vector to be supplied

vlogical(
  vector,
  vpattern,
  test = stringi::stri_detect_regex,
  simplify_with = NULL,
  ...
)

Arguments

vector

A vector or dimensional object to be processed (e.g., matrix, data.frame, etc.)

vpattern

A named vector of patterns to be matched

test

(function) The function to use for logical testing: the function should be appropriate for the values of vpattern accepting vector as the first argument, the pattern as the second, and '...' as the third (even if not used).

simplify_with

(function) A function to column-wise simplify the logical matrix when not empty

...

Additional arguments to be sent to the function held by argument test

Value

A logical matrix, with rows of the same length as vector and columns the length of vpattern when simplify_with is not provided. Otherwise, the result of applying simplify_with on each column.

See also

Examples

vlogical(
  vector = letters
  , vpattern = c(sample(LETTERS, 5), sample(letters, 5))
  , simplify_with = mean
  , ignore.case = TRUE
  )
#>        x_1        x_2        x_3        x_4        x_5        x_6        x_7 
#> 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.03846154 0.03846154 
#>        x_8        x_9       x_10 
#> 0.03846154 0.03846154 0.03846154