Vectorized Logical Tests

vlike is a vectorized version of like allowing a pattern vector to be supplied

vlogical(vector, vpattern, test, simplify_with = NULL, ...)

Arguments

vector

A vector or dimensional object to be processed (e.g., matrix, data.frame, etc.)

vpattern

A vector of patterns to be matched

test

(function) The function to use for logical testing: the function should be appropriate for the values of vpattern and have a named argument of vector, an additional argument to accept the pattern fed atomically, and '...' (even if not used).

simplify_with

(function) When provided, the function operates over rows, thereby simplifying the result

...

Additional arguments to be sent to the function held by argument test

Value

A logical matrix, with rows of the same length as vector and columns the length of vpattern TRUE for items that match pattern. If vector is dimensional, multiple TRUE values may be found for each row.

See also